+++ src/htsmodel.cpp
@@ -482,7 +482,7 @@ bool HTSModel::initializeElements(std::list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-  for(size_t i = 0; i < m_elements.size(); i++)
+  for(int i = 0; i < (int)m_elements.size(); i++)
   {
     Element *element = m_elements[i];
     element->index = i;
